move files to folder dl models
